+++ test/app.js
@@ -1,12 +1,24 @@
 
-var express = require('../')
-  , assert = require('assert');
+var assert = require('assert')
+var express = require('..')
+var request = require('supertest')
 
 describe('app', function(){
   it('should inherit from event emitter', function(done){
     var app = express();
     app.on('foo', done);
     app.emit('foo');
+  })
+
+  it('should be callable', function(){
+    var app = express();
+    assert(typeof app, 'function');
+  })
+
+  it('should 500 without routes', function(done){
+    request(express())
+    .get('/')
+    .expect(500, done);
   })
 })