+++ History.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+ * fix behavior when handling request without routes
  * use `finalhandler` for final response handling
  * use `vary` module for `res.vary`
  * deps: proxy-addr@1.0.1